#  Cache the span by using BufferWriter<T> within JsonWriter<T>