--- a/src/System.Buffers.ReaderWriter/System/Buffers/Writer/BufferWriterT.cs
+++ b/src/System.Buffers.ReaderWriter/System/Buffers/Writer/BufferWriterT.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Runtime.CompilerServices;
+using System.Text;
 
 namespace System.Buffers.Writer
 {
@@ -12,17 +13,17 @@ namespace System.Buffers.Writer
         private Span<byte> _span;
         private int _buffered;
 
-        private static readonly byte[] s_newLine = new byte[] { (byte)'\r', (byte)'\n' };
+        private static readonly byte[] s_newLine = Encoding.UTF8.GetBytes(Environment.NewLine);
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public BufferWriter(T output)
         {
             _buffered = 0;
             _output = output;
             _span = output.GetSpan();
-            NewLine = s_newLine;
         }
 
-        public ReadOnlySpan<byte> NewLine { get; set; }
+        private static ReadOnlySpan<byte> NewLine => s_newLine;
 
         public Span<byte> Buffer => _span;