--- a/tests/Benchmarks/System.Text.JsonLab/JsonWriterPerf.cs
+++ b/tests/Benchmarks/System.Text.JsonLab/JsonWriterPerf.cs
@@ -9,24 +9,24 @@ using System.Text.Formatting;
 
 namespace System.Text.JsonLab.Benchmarks
 {
-    [SimpleJob(-1, 5, 10, 32768)]
+    //[SimpleJob(-1, 5, 10, 32768)]
     [MemoryDiagnoser]
     public class JsonWriterPerf
     {
         private const int ExtraArraySize = 1000;
         private const int BufferSize = 1024 + (ExtraArraySize * 64);
 
-        private ArrayFormatter _arrayFormatter;
+        private ArrayFormatterWrapper _ArrayFormatterWrapper;
         private MemoryStream _memoryStream;
         private StreamWriter _streamWriter;
         private StringBuilder _stringBuilder;
 
         private int[] _data;
 
-        [Params(true, false)]
+        [Params(true)]
         public bool IsUTF8Encoded;
 
-        [Params(true, false)]
+        [Params(false)]
         public bool Formatted;
 
         [GlobalSetup]
@@ -45,26 +45,26 @@ namespace System.Text.JsonLab.Benchmarks
                 var buffer = new byte[BufferSize];
                 _memoryStream = new MemoryStream(buffer);
                 _streamWriter = new StreamWriter(_memoryStream, new UTF8Encoding(false), BufferSize, true);
-                _arrayFormatter = new ArrayFormatter(BufferSize, SymbolTable.InvariantUtf8);
+                _ArrayFormatterWrapper = new ArrayFormatterWrapper(BufferSize, SymbolTable.InvariantUtf8);
             }
             else
             {
                 _stringBuilder = new StringBuilder();
-                _arrayFormatter = new ArrayFormatter(BufferSize, SymbolTable.InvariantUtf16);
+                _ArrayFormatterWrapper = new ArrayFormatterWrapper(BufferSize, SymbolTable.InvariantUtf16);
             }
         }
 
         [Benchmark]
         public void WriterSystemTextJsonBasic()
         {
-            _arrayFormatter.Clear();
+            _ArrayFormatterWrapper.Clear();
             if (IsUTF8Encoded)
-                WriterSystemTextJsonBasicUtf8(Formatted, _arrayFormatter, _data.AsSpan(0, 10));
+                WriterSystemTextJsonBasicUtf8(Formatted, _ArrayFormatterWrapper, _data.AsSpan(0, 10));
             else
-                WriterSystemTextJsonBasicUtf16(Formatted, _arrayFormatter, _data.AsSpan(0, 10));
+                WriterSystemTextJsonBasicUtf16(Formatted, _ArrayFormatterWrapper, _data.AsSpan(0, 10));
         }
 
-        [Benchmark]
+        //[Benchmark]
         public void WriterNewtonsoftBasic()
         {
             WriterNewtonsoftBasic(Formatted, GetWriter(), _data.AsSpan(0, 10));
@@ -73,14 +73,14 @@ namespace System.Text.JsonLab.Benchmarks
         [Benchmark]
         public void WriterSystemTextJsonHelloWorld()
         {
-            _arrayFormatter.Clear();
+            _ArrayFormatterWrapper.Clear();
             if (IsUTF8Encoded)
-                WriterSystemTextJsonHelloWorldUtf8(Formatted, _arrayFormatter);
+                WriterSystemTextJsonHelloWorldUtf8(Formatted, _ArrayFormatterWrapper);
             else
-                WriterSystemTextJsonHelloWorldUtf16(Formatted, _arrayFormatter);
+                WriterSystemTextJsonHelloWorldUtf16(Formatted, _ArrayFormatterWrapper);
         }
 
-        [Benchmark]
+        //[Benchmark]
         public void WriterNewtonsoftHelloWorld()
         {
             WriterNewtonsoftHelloWorld(Formatted, GetWriter());
@@ -88,18 +88,18 @@ namespace System.Text.JsonLab.Benchmarks
 
         [Benchmark]
         [Arguments(1)]
-        [Arguments(2)]
-        [Arguments(5)]
+        //[Arguments(2)]
+        //[Arguments(5)]
         [Arguments(10)]
-        [Arguments(100)]
+        //[Arguments(100)]
         [Arguments(1000)]
         public void WriterSystemTextJsonArrayOnly(int size)
         {
-            _arrayFormatter.Clear();
+            _ArrayFormatterWrapper.Clear();
             if (IsUTF8Encoded)
-                WriterSystemTextJsonArrayOnlyUtf8(Formatted, _arrayFormatter, _data.AsSpan(0, size));
+                WriterSystemTextJsonArrayOnlyUtf8(Formatted, _ArrayFormatterWrapper, _data.AsSpan(0, size));
             else
-                WriterSystemTextJsonArrayOnlyUtf16(Formatted, _arrayFormatter, _data.AsSpan(0, size));
+                WriterSystemTextJsonArrayOnlyUtf16(Formatted, _ArrayFormatterWrapper, _data.AsSpan(0, size));
         }
 
         private TextWriter GetWriter()
@@ -118,9 +118,9 @@ namespace System.Text.JsonLab.Benchmarks
             return writer;
         }
 
-        private static void WriterSystemTextJsonBasicUtf8(bool formatted, ArrayFormatter output, ReadOnlySpan<int> data)
+        private static void WriterSystemTextJsonBasicUtf8(bool formatted, ArrayFormatterWrapper output, ReadOnlySpan<int> data)
         {
-            var json = new JsonWriter(output, true, formatted);
+            var json = new JsonWriter<ArrayFormatterWrapper>(output, true, formatted);
 
             json.WriteObjectStart();
             json.WriteAttribute("age", 42);
@@ -146,9 +146,9 @@ namespace System.Text.JsonLab.Benchmarks
             json.WriteObjectEnd();
         }
 
-        private static void WriterSystemTextJsonBasicUtf16(bool formatted, ArrayFormatter output, ReadOnlySpan<int> data)
+        private static void WriterSystemTextJsonBasicUtf16(bool formatted, ArrayFormatterWrapper output, ReadOnlySpan<int> data)
         {
-            var json = new JsonWriter(output, false, formatted);
+            var json = new JsonWriter<ArrayFormatterWrapper>(output, false, formatted);
 
             json.WriteObjectStart();
             json.WriteAttribute("age", 42);
@@ -214,18 +214,18 @@ namespace System.Text.JsonLab.Benchmarks
             }
         }
 
-        private static void WriterSystemTextJsonHelloWorldUtf8(bool formatted, ArrayFormatter output)
+        private static void WriterSystemTextJsonHelloWorldUtf8(bool formatted, ArrayFormatterWrapper output)
         {
-            var json = new JsonWriter(output, true, formatted);
+            var json = new JsonWriter<ArrayFormatterWrapper>(output, true, formatted);
 
             json.WriteObjectStart();
             json.WriteAttribute("message", "Hello, World!");
             json.WriteObjectEnd();
         }
 
-        private static void WriterSystemTextJsonHelloWorldUtf16(bool formatted, ArrayFormatter output)
+        private static void WriterSystemTextJsonHelloWorldUtf16(bool formatted, ArrayFormatterWrapper output)
         {
-            var json = new JsonWriter(output, false, formatted);
+            var json = new JsonWriter<ArrayFormatterWrapper>(output, false, formatted);
 
             json.WriteObjectStart();
             json.WriteAttribute("message", "Hello, World!");
@@ -245,9 +245,9 @@ namespace System.Text.JsonLab.Benchmarks
             }
         }
 
-        private static void WriterSystemTextJsonArrayOnlyUtf8(bool formatted, ArrayFormatter output, ReadOnlySpan<int> data)
+        private static void WriterSystemTextJsonArrayOnlyUtf8(bool formatted, ArrayFormatterWrapper output, ReadOnlySpan<int> data)
         {
-            var json = new JsonWriter(output, true, formatted);
+            var json = new JsonWriter<ArrayFormatterWrapper>(output, true, formatted);
 
             json.WriteArrayStart("ExtraArray");
             for (var i = 0; i < data.Length; i++)
@@ -257,9 +257,9 @@ namespace System.Text.JsonLab.Benchmarks
             json.WriteArrayEnd();
         }
 
-        private static void WriterSystemTextJsonArrayOnlyUtf16(bool formatted, ArrayFormatter output, ReadOnlySpan<int> data)
+        private static void WriterSystemTextJsonArrayOnlyUtf16(bool formatted, ArrayFormatterWrapper output, ReadOnlySpan<int> data)
         {
-            var json = new JsonWriter(output, false, formatted);
+            var json = new JsonWriter<ArrayFormatterWrapper>(output, false, formatted);
 
             json.WriteArrayStart("ExtraArray");
             for (var i = 0; i < data.Length; i++)