--- a/tests/Benchmarks/System.Text.JsonLab/JsonWriterPerf.cs
+++ b/tests/Benchmarks/System.Text.JsonLab/JsonWriterPerf.cs
@@ -16,7 +16,7 @@ namespace System.Text.JsonLab.Benchmarks
         private const int ExtraArraySize = 1000;
         private const int BufferSize = 1024 + (ExtraArraySize * 64);
 
-        private ArrayFormatterWrapper _ArrayFormatterWrapper;
+        private ArrayFormatterWrapper _arrayFormatterWrapper;
         private MemoryStream _memoryStream;
         private StreamWriter _streamWriter;
         private StringBuilder _stringBuilder;
@@ -45,23 +45,23 @@ namespace System.Text.JsonLab.Benchmarks
                 var buffer = new byte[BufferSize];
                 _memoryStream = new MemoryStream(buffer);
                 _streamWriter = new StreamWriter(_memoryStream, new UTF8Encoding(false), BufferSize, true);
-                _ArrayFormatterWrapper = new ArrayFormatterWrapper(BufferSize, SymbolTable.InvariantUtf8);
+                _arrayFormatterWrapper = new ArrayFormatterWrapper(BufferSize, SymbolTable.InvariantUtf8);
             }
             else
             {
                 _stringBuilder = new StringBuilder();
-                _ArrayFormatterWrapper = new ArrayFormatterWrapper(BufferSize, SymbolTable.InvariantUtf16);
+                _arrayFormatterWrapper = new ArrayFormatterWrapper(BufferSize, SymbolTable.InvariantUtf16);
             }
         }
 
         [Benchmark]
         public void WriterSystemTextJsonBasic()
         {
-            _ArrayFormatterWrapper.Clear();
+            _arrayFormatterWrapper.Clear();
             if (IsUTF8Encoded)
-                WriterSystemTextJsonBasicUtf8(Formatted, _ArrayFormatterWrapper, _data.AsSpan(0, 10));
+                WriterSystemTextJsonBasicUtf8(Formatted, _arrayFormatterWrapper, _data.AsSpan(0, 10));
             else
-                WriterSystemTextJsonBasicUtf16(Formatted, _ArrayFormatterWrapper, _data.AsSpan(0, 10));
+                WriterSystemTextJsonBasicUtf16(Formatted, _arrayFormatterWrapper, _data.AsSpan(0, 10));
         }
 
         //[Benchmark]
@@ -73,11 +73,11 @@ namespace System.Text.JsonLab.Benchmarks
         [Benchmark]
         public void WriterSystemTextJsonHelloWorld()
         {
-            _ArrayFormatterWrapper.Clear();
+            _arrayFormatterWrapper.Clear();
             if (IsUTF8Encoded)
-                WriterSystemTextJsonHelloWorldUtf8(Formatted, _ArrayFormatterWrapper);
+                WriterSystemTextJsonHelloWorldUtf8(Formatted, _arrayFormatterWrapper);
             else
-                WriterSystemTextJsonHelloWorldUtf16(Formatted, _ArrayFormatterWrapper);
+                WriterSystemTextJsonHelloWorldUtf16(Formatted, _arrayFormatterWrapper);
         }
 
         //[Benchmark]
@@ -95,11 +95,11 @@ namespace System.Text.JsonLab.Benchmarks
         [Arguments(1000)]
         public void WriterSystemTextJsonArrayOnly(int size)
         {
-            _ArrayFormatterWrapper.Clear();
+            _arrayFormatterWrapper.Clear();
             if (IsUTF8Encoded)
-                WriterSystemTextJsonArrayOnlyUtf8(Formatted, _ArrayFormatterWrapper, _data.AsSpan(0, size));
+                WriterSystemTextJsonArrayOnlyUtf8(Formatted, _arrayFormatterWrapper, _data.AsSpan(0, size));
             else
-                WriterSystemTextJsonArrayOnlyUtf16(Formatted, _ArrayFormatterWrapper, _data.AsSpan(0, size));
+                WriterSystemTextJsonArrayOnlyUtf16(Formatted, _arrayFormatterWrapper, _data.AsSpan(0, size));
         }
 
         private TextWriter GetWriter()