--- a/tests/Benchmarks/System.Text.JsonLab/JsonWriterPerf.cs
+++ b/tests/Benchmarks/System.Text.JsonLab/JsonWriterPerf.cs
@@ -9,7 +9,7 @@ using System.Text.Formatting;
 
 namespace System.Text.JsonLab.Benchmarks
 {
-    //[SimpleJob(-1, 5, 10, 32768)]
+    [SimpleJob(-1, 5, 10, 32768)]
     [MemoryDiagnoser]
     public class JsonWriterPerf
     {
@@ -23,10 +23,10 @@ namespace System.Text.JsonLab.Benchmarks
 
         private int[] _data;
 
-        [Params(true)]
+        [Params(true, false)]
         public bool IsUTF8Encoded;
 
-        [Params(false)]
+        [Params(true, false)]
         public bool Formatted;
 
         [GlobalSetup]
@@ -64,7 +64,7 @@ namespace System.Text.JsonLab.Benchmarks
                 WriterSystemTextJsonBasicUtf16(Formatted, _arrayFormatterWrapper, _data.AsSpan(0, 10));
         }
 
-        //[Benchmark]
+        [Benchmark]
         public void WriterNewtonsoftBasic()
         {
             WriterNewtonsoftBasic(Formatted, GetWriter(), _data.AsSpan(0, 10));
@@ -80,7 +80,7 @@ namespace System.Text.JsonLab.Benchmarks
                 WriterSystemTextJsonHelloWorldUtf16(Formatted, _arrayFormatterWrapper);
         }
 
-        //[Benchmark]
+        [Benchmark]
         public void WriterNewtonsoftHelloWorld()
         {
             WriterNewtonsoftHelloWorld(Formatted, GetWriter());
@@ -88,10 +88,10 @@ namespace System.Text.JsonLab.Benchmarks
 
         [Benchmark]
         [Arguments(1)]
-        //[Arguments(2)]
-        //[Arguments(5)]
+        [Arguments(2)]
+        [Arguments(5)]
         [Arguments(10)]
-        //[Arguments(100)]
+        [Arguments(100)]
         [Arguments(1000)]
         public void WriterSystemTextJsonArrayOnly(int size)
         {